backend/requirements.txt
# Jarvis MCP Backend Dependencies
# Core FastAPI Stack
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database & ORM
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.4

# AI Clients & HTTP
httpx==0.25.2
openai==1.3.7
anthropic==0.7.8
ollama==0.1.7

# Data Processing & Embeddings
numpy==1.24.4
pandas==2.1.4
sentence-transformers==2.2.2
tiktoken==0.5.2

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Security & Utils
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4